GH-1: Organized files and improved code documentation.
