--- ResgateIO.Service/ICollectionHandler.cs
@@ -1,7 +1,0 @@
-﻿namespace ResgateIO.Service
-{
-    public interface ICollectionHandler : IResourceHandler
-    {
-        void Get(ICollectionRequest request);
-    }
-}
