--- ResgateIO.Service/IModelHandler.cs
@@ -1,7 +1,0 @@
-﻿namespace ResgateIO.Service
-{
-    public interface IModelHandler : IResourceHandler
-    {
-        void Get(IModelRequest request);
-    }
-}
