+++ ResgateIO.Service/ResError.cs
@@ -3,6 +3,10 @@ using Newtonsoft.Json;
 
 namespace ResgateIO.Service
 {
+    /// <summary>
+    /// Represents a RES service error.
+    /// </summary>
+    /// <see>https://resgate.io/docs/specification/res-service-protocol/#error-object</see>
     public class ResError
     {
         // Predefined error codes
@@ -22,12 +26,20 @@ namespace ResgateIO.Service
         [JsonProperty(PropertyName = "data", NullValueHandling = NullValueHandling.Ignore)]
         public object Data = null;
 
+        /// <summary>
+        /// Initializes a new instance of the ResError class with CodeInternalError.
+        /// </summary>
+        /// <param name="ex">Exception to base the error message on.</param>
         public ResError(Exception ex)
         {
             Code = ResError.CodeInternalError;
             Message = "Internal error: " + ex.Message;
         }
 
+        /// <summary>
+        /// Initializes a new instance of the ResError class based on a ResException.
+        /// </summary>
+        /// <param name="ex">ResException to get the error code, message, and data from.</param>
         public ResError(ResException ex)
         {
             Code = ex.Code;
@@ -35,12 +47,24 @@ namespace ResgateIO.Service
             Data = ex.ErrorData;
         }
 
+        /// <summary>
+        /// Initializes a new instance of the ResError class with custom error code and message.
+        /// </summary>
+        /// <param name="code">Error code.</param>
+        /// <param name="message">Error message.</param>
         public ResError(string code, string message)
         {
             Code = code;
             Message = message;
         }
 
+
+        /// <summary>
+        /// Initializes a new instance of the ResError class with custom error code, message, and data.
+        /// </summary>
+        /// <param name="code">Error code.</param>
+        /// <param name="message">Error message.</param>
+        /// <param name="data">Additional data. Must be JSON serializable.</param>
         public ResError(string code, string message, object data)
         {
             Code = code;
