+++ ResgateIO.Service/logger/VoidLogger.cs
