+++ ResgateIO.Service/request/Request.cs
@@ -7,6 +7,9 @@ using Newtonsoft.Json.Linq;
 
 namespace ResgateIO.Service
 {
+    /// <summary>
+    /// Provides context information and methods for responding to a request.
+    /// </summary>
     public class Request: ResourceContext, IAccessRequest, IGetRequest, ICallRequest, IAuthRequest, IModelRequest, ICollectionRequest
     {
         private readonly Msg msg;
